requirements.txt
pythonqwt
pyqt

